refactor: refactor code for better readability and remove tech debt
