+++ src/components/Rating.tsx
@@ -1,5 +1,4 @@
-import React from 'react';
-import { useState } from 'react';
+import { FC, useState } from 'react';
 import { InputLabel } from '@mui/material';
 
 interface RatingProps {
@@ -8,12 +7,14 @@ interface RatingProps {
   updateValue: (newValue: number) => void;
 }
 
-export const Rating: React.FC<RatingProps> = ({ id, value, updateValue }) => {
+export const Rating: FC<RatingProps> = ({ id, value, updateValue }) => {
   const [hoverAt, setHoverAt] = useState<number | null>(null);
 
   return (
     <div id='#/properties/rating' className='rating'>
-      <InputLabel shrink style={{ marginTop: '0.8em' }}>Rating</InputLabel>
+      <InputLabel shrink style={{ marginTop: '0.8em' }}>
+        Rating
+      </InputLabel>
       <div style={{ cursor: 'pointer', fontSize: '18px' }}>
         {[0, 1, 2, 3, 4].map((i) => {
           const fullStars = hoverAt ?? value;
