+++ src/components/RatingControl.tsx
