+++ src/main.tsx
@@ -1,6 +1,7 @@
 import { createTheme, CssBaseline, ThemeProvider } from '@mui/material';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import App from './App';
+import { StrictMode } from 'react';
 
 /**
  * Customize form so each control has more space
@@ -12,15 +13,20 @@ const theme = createTheme({
         root: {
           margin: '0.8em 0',
         },
-      }
+      },
     },
   },
 });
 
-ReactDOM.render(
-  <ThemeProvider theme={theme}>
-    <CssBaseline />
-    <App />
-  </ThemeProvider>,
-  document.getElementById('root')
+const rootEl = document.getElementById('root');
+
+if (!rootEl) throw new Error('Failed to find the root element');
+
+createRoot(rootEl).render(
+  <StrictMode>
+    <ThemeProvider theme={theme}>
+      <CssBaseline />
+      <App />
+    </ThemeProvider>
+  </StrictMode>
 );
